lab12python/requirements.txt
django==3.2.3
djangorestframework==3.11.1
djangorestframework-simplejwt==5.2.0
psycopg2-binary